refactor: move createPriceEvent to its own file
