+++ src/game-logic/reducers/createPriceEvent.test.js
@@ -1,0 +1,41 @@
+import { sampleCropItem1 } from '../../data/items'
+
+import { createPriceEvent } from './createPriceEvent'
+
+jest.mock('../../data/items')
+
+describe('createPriceEvent', () => {
+  test('creates priceCrashes data', () => {
+    const priceEvent = {
+      itemId: sampleCropItem1.id,
+      daysRemaining: 1,
+    }
+
+    const { priceCrashes } = createPriceEvent(
+      { priceCrashes: {} },
+      priceEvent,
+      'priceCrashes'
+    )
+
+    expect(priceCrashes).toMatchObject({
+      [sampleCropItem1.id]: priceEvent,
+    })
+  })
+
+  test('creates priceSurges data', () => {
+    const priceEvent = {
+      itemId: sampleCropItem1.id,
+      daysRemaining: 1,
+    }
+
+    const { priceSurges } = createPriceEvent(
+      { priceSurges: {} },
+      priceEvent,
+      'priceSurges'
+    )
+
+    expect(priceSurges).toMatchObject({
+      [sampleCropItem1.id]: priceEvent,
+    })
+  })
+})
